fix: change BloodyFlaskItem.amount from float to double to match IBloodVolume API
Agent-Logs-Url: https://github.com/VincentHuto/Hemomancy/sessions/69dc1bc1-1a99-4112-82ca-97ab8197e534

+++ src/main/java/com/vincenthuto/hemomancy/common/item/BloodyFlaskItem.java
@@ -22,9 +22,9 @@ import net.minecraftforge.network.PacketDistributor;
 
 public class BloodyFlaskItem extends Item {
 
-	float amount;
+	double amount;
 
-	public BloodyFlaskItem(Properties prop, float amount) {
+	public BloodyFlaskItem(Properties prop, double amount) {
 		super(prop.stacksTo(16));
 		this.amount = amount;
 	}
@@ -35,7 +35,7 @@ public class BloodyFlaskItem extends Item {
 		tooltip.add(Component.literal("Used to Quickly Gain " + amount + "ml of Blood"));
 	}
 
-	public float getAmount() {
+	public double getAmount() {
 		return amount;
 	}
